Introduce System Properties to configure ES host and port
This change introduces two properties to configure the Elastic Search
host and port to which this data project attempt to connect to.

This allows changing the host and port of the ES instance at runtime
without having to make changes in the configuration files and rebuilding
the project.

The property names themselves align with the pattern seen for system
properties in configuration files of address-index-api (for example:
server/conf/application.conf)