Update data project to allow running against a local Elasticsearch instance
The csv names containing the data are currently named read_test.csv, but
when uk.gov.ons.addressindex.Main is run, it attempts to validate the
filenams and searchec for the following patterns to extract the EPOCH
and DATE:

  val epochRegex = s"ABP_E(\\d+).+$$".r
  val dateRegex = s"ABP_E.+(\\d{6})\\.csv$$".r

On each run of the data project, a unique index is generated. The name
of this unique index is derived from the epoch and date extracted below.
If the file names do not satisfy the above patterns, the index creation
in local Elasticsearch fails.

This commit amends the csv names to fit the patterns above and are then
successfully used to create and populated an index in the local ES